cmd/clone.go(CloneAllRepos): Optionally prune removed repos.
This feature, enabled with `GHORG_PRUNE` or `--prune`, tells ghorg to remove any clones in an org or
user's folder (e.g., descendants of `~/ghorg/someuser`) that are not found in the remote listing of
target clones.  For example, if a repo has been deleted on Github, the `--prune` directive will
allow the user to keep their local org-clone in sync with what has been added *and* removed on
Github.